requirements.txt
face_recognition
opencv-python
numpy
pyserial
requests
flask
